fix entry_pop in virtraft
previously, entry_pop never ran during our virtraft tests, probably because without removing leader, we never have a situation where this is neccessary

i.e. the only case virtraft can pop entries, is when we have a leader lose leadership and we have an entry pending and not committed.  so when another node becomes leader, old leader's entries get popped.

however, that has nothing to do with the "id" of the node that its effecting so trying to contorl it through that, makes no sense.