prevented a non voting leader from accepting entries
a non voting leader (i.e. demoted itself), shouldn't be able to recv new entries.  A similar case would be when its bootstrapping a cluster, as it would want to add itself, therefore we disallow it if the number of nodes is > 1.  If the demoted leader is in a cluster by itself, it can recv_entry()

this required a fix to tests to make sure the server had a node object created for the cluster, so it can accept an entry