add a read_queue test to virtraft
every iteration we push a read_queue request and the handler and "arg" we pass to it set a variable on calling.  we can use this to make sure that the read_queue doesn't get too far from the iteration.

this is analagoues to the current deadlock test.

in the future, we will also test for correctness, that the other nodes confirm to virtraft that they have gotten the msgid that corresponds to allowing this