don't pipeline appendentries to unacknowledged followers
when we become leader, we don't know the state of the other nodes, so we set their next_idx to what would be our next_idx.

this enables us to send an emoty appendentries to the other nodes on become leader where we will get a response back to their existing position

however, we then can call send_appendentries again before we get a response, so we be resetting node state on the second one in an inconsistent manner

therefore, dont send another appendentries to a followr until we get a response.

this assumes, reliable medium (i.e. if we dont get a response for the first, we wont get a response for the second or further).  so worse case scenario, the non responding node will time out and try to cause an election.