implements msg_id checking in virtraft for verification of read_queue requests
When we pop an entry off the read_queue with the can_read flag, we verify across all nodes that the majority of nodes have accepted from the leader a msg_id past what this read_queue entry needs.

the problem with this is, that until now, msg_id was private to each server instance, not cluster wise, so it wasnt always increasing across the cluster.  we change it here, so that

a) followrs will increase their msg_id to synchronize with leade
b) followers will return their old msg_id to leader in their response, so it can increase its own, if its not greater.

using that, we can also add a raft_get_voting_node_ids() private function that returns a malloced (has to be freed by caller) array of raft_node_id_t's that testers can use to enumerate an inspect a nodes followers.  In this case, for getting their current msg_id, which we also enable by providing another new private function raft_get_msg_id()