Add snapshot_index to raft_snapshot_resp_t struct (#163)
Normally, receiver side has checks if the chunk is for the correct 
index and offset. 

Adding snapshot_index to raft_snapshot_resp_t just to be sure we don't 
update snapshot offset for the node unnecessarily.

Also, it makes debugging easier as we know the snapshot index that 
response refers to.